+++ src/direction.rs
@@ -1,7 +1,7 @@
 #[cfg(feature = "serde")]
 use serde::{Deserialize, Serialize};
 
-use bevy_math::IVec2;
+use bevy_math::{IVec2, UVec2};
 
 pub const NORTH: IVec2 = IVec2 { x: 0, y: 1 };
 pub const NORTH_EAST: IVec2 = IVec2 { x: 1, y: 1 };
@@ -46,8 +46,8 @@ impl Direction {
     /// Move a point in this direction by the given amount.
     #[inline]
     #[must_use]
-    pub fn move_point(&self, point: IVec2, by: i32) -> IVec2 {
-        point + self.unit() * by
+    pub fn move_point(&self, point: UVec2, by: u32) -> UVec2 {
+        (point.as_ivec2() + self.unit() * by as i32).as_uvec2()
     }
 
     /// Returns true if this direction is cardinal (N, E, S, W).