+++ src/lib.rs
@@ -9,6 +9,7 @@
 //! pixels having the same value (as per the function of a quad tree).
 
 mod direction;
+mod math;
 mod node_path;
 mod pixel_map;
 mod pnode;
@@ -18,6 +19,6 @@ mod region;
 mod shapes;
 
 pub use self::{
-    direction::*, node_path::*, pixel_map::*, pnode::*, quadrant::*, ray_cast::*, region::*,
-    shapes::*,
+    direction::*, math::*, node_path::*, pixel_map::*, pnode::*, quadrant::*, ray_cast::*,
+    region::*, shapes::*,
 };