+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   dimensions, and ensures the map space is always (0, 0) at the bottom left corner.
 * Add `PixelMap::map_size` to store a logical size for the map that is not restrained by the requirements
   of the underlying node implementation, which requires dimensions that are square and powers of two.
+* Refactor shapes to be represented in unsigned space.
 
 ## v0.2.0